* Separared MessageSystem and Webservices into different projects.Added UnitTest project. Excluded compiled output form repository.
