--- EF_Messages.sln
@@ -1,24 +1,0 @@
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.5.2.0
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EF_Messages", "EF_Messages.csproj", "{22F5396B-8A40-2AA4-0C4B-85874B8211C1}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{22F5396B-8A40-2AA4-0C4B-85874B8211C1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{22F5396B-8A40-2AA4-0C4B-85874B8211C1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{22F5396B-8A40-2AA4-0C4B-85874B8211C1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{22F5396B-8A40-2AA4-0C4B-85874B8211C1}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {1D357E83-665E-4082-9C28-6BDCC22748B7}
-	EndGlobalSection
-EndGlobal
