+++ MessageSystem/model/Message.cs
