+++ WebServices/WebServices.http
@@ -1,0 +1,6 @@
+@WebServices_HostAddress = http://localhost:5063
+
+GET {{WebServices_HostAddress}}/weatherforecast/
+Accept: application/json
+
+###
