+++ WebServices/appsettings.json
@@ -1,0 +1,12 @@
+{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft.AspNetCore": "Warning"
+    }
+  },
+  "AllowedHosts": "*",
+    "ConnectionStrings": {
+      "MessageSystemConnection": "Server=(localdb)\\mssqllocaldb;Database=MessageSystemDb;Trusted_Connection=True;"
+  }
+}
